feat: extract shared rule modules and oxlint rule mapping
Move the rule maps that both linters consume into src/rules/ so a single
edit updates the ESLint and oxlint factories. Adds the explicit per-rule
oxlint mapping (native / tsgolint / js-plugin, with documented
stays-in-eslint reasons) that drives hybrid mode.
