+++ src/eslint/configs/eslint-plugin.ts
@@ -1,4 +1,5 @@
 import { GLOB_SRC } from "../../globs.ts";
+import { eslintPluginRules } from "../../rules/eslint-plugin.ts";
 import { ensurePackages, interopDefault } from "../../utils.ts";
 import type { OptionsFiles, OptionsOverrides, TypedFlatConfigItem } from "../types.ts";
 
@@ -21,47 +22,7 @@ export async function eslintPlugin({
 			name: "isentinel/eslint-plugin/rules",
 			files,
 			rules: {
-				"eslint-plugin/consistent-output": "error",
-				"eslint-plugin/fixer-return": "error",
-				"eslint-plugin/meta-property-ordering": "off",
-				"eslint-plugin/no-deprecated-context-methods": "error",
-				"eslint-plugin/no-deprecated-report-api": "error",
-				"eslint-plugin/no-identical-tests": "error",
-				"eslint-plugin/no-matching-violation-suggest-message-ids": "error",
-				"eslint-plugin/no-meta-replaced-by": "error",
-				"eslint-plugin/no-meta-schema-default": "error",
-				"eslint-plugin/no-missing-message-ids": "error",
-				"eslint-plugin/no-missing-placeholders": "error",
-				"eslint-plugin/no-only-tests": "error",
-				"eslint-plugin/no-property-in-node": "error",
-				"eslint-plugin/no-unused-message-ids": "error",
-				"eslint-plugin/no-unused-placeholders": "error",
-				"eslint-plugin/no-useless-token-range": "error",
-				"eslint-plugin/prefer-message-ids": "error",
-				"eslint-plugin/prefer-object-rule": "error",
-				"eslint-plugin/prefer-output-null": "error",
-				"eslint-plugin/prefer-placeholders": "error",
-				"eslint-plugin/prefer-replace-text": "error",
-				"eslint-plugin/report-message-format": "error",
-				"eslint-plugin/require-meta-default-options": "error",
-				"eslint-plugin/require-meta-docs-description": [
-					"error",
-					{
-						pattern: "^(Enforce|Require|Disallow).*[^.!]$",
-					},
-				],
-				"eslint-plugin/require-meta-docs-recommended": "error",
-				"eslint-plugin/require-meta-docs-url": "off",
-				"eslint-plugin/require-meta-fixable": "error",
-				"eslint-plugin/require-meta-has-suggestions": "error",
-				"eslint-plugin/require-meta-schema": "error",
-				"eslint-plugin/require-meta-schema-description": "error",
-				"eslint-plugin/require-meta-type": "error",
-				"eslint-plugin/require-test-case-name": "error",
-				"eslint-plugin/require-test-error-positions": "error",
-				"eslint-plugin/test-case-property-ordering": "error",
-				"eslint-plugin/test-case-shorthand-strings": "error",
-				"eslint-plugin/unique-test-case-names": "error",
+				...eslintPluginRules(),
 
 				...overrides,
 			},
