+++ src/eslint/configs/node.ts
@@ -1,3 +1,4 @@
+import { nodeRules } from "../../rules/node.ts";
 import { ensurePackages, interopDefault } from "../../utils.ts";
 import type { TypedFlatConfigItem } from "../types.ts";
 
@@ -12,17 +13,7 @@ export async function node(): Promise<Array<TypedFlatConfigItem>> {
 			plugins: {
 				node: pluginNode,
 			},
-			rules: {
-				"node/handle-callback-err": ["error", "^(err|error)$"],
-				"node/no-deprecated-api": "error",
-				"node/no-exports-assign": "error",
-				"node/no-new-require": "error",
-				"node/no-path-concat": "error",
-				"node/prefer-global/buffer": ["error", "never"],
-				"node/prefer-global/process": ["error", "never"],
-				"node/prefer-node-protocol": "error",
-				"node/process-exit-as-throw": "error",
-			},
+			rules: nodeRules(),
 		},
 	];
 }
