+++ src/eslint/configs/promise.ts
@@ -1,3 +1,4 @@
+import { promiseRules } from "../../rules/promise.ts";
 import { interopDefault } from "../../utils.ts";
 import type { TypedFlatConfigItem } from "../types.ts";
 
@@ -11,33 +12,7 @@ export async function promise(): Promise<Array<TypedFlatConfigItem>> {
 			plugins: {
 				promise: pluginPromise,
 			},
-			rules: {
-				"promise/always-return": [
-					"error",
-					{
-						ignoreLastCallback: true,
-					},
-				],
-				"promise/avoid-new": "off",
-				"promise/catch-or-return": [
-					"error",
-					{
-						allowFinally: true,
-						allowThen: true,
-					},
-				],
-				"promise/no-callback-in-promise": "off",
-				"promise/no-multiple-resolved": "error",
-				"promise/no-native": "off",
-				"promise/no-nesting": "warn",
-				"promise/no-new-statics": "off",
-				"promise/no-promise-in-callback": "warn",
-				"promise/no-return-in-finally": "warn",
-				"promise/no-return-wrap": "error",
-				"promise/param-names": "warn",
-				"promise/prefer-await-to-callbacks": "off",
-				"promise/prefer-await-to-then": "off",
-			},
+			rules: promiseRules(),
 		},
 	];
 }
