+++ src/eslint/configs/spelling.ts
@@ -1,7 +1,5 @@
-import { createRequire } from "node:module";
-import { pathToFileURL } from "node:url";
-
 import { GLOB_SRC } from "../../globs.ts";
+import { spellingRules } from "../../rules/spelling.ts";
 import { interopDefault } from "../../utils.ts";
 import type {
 	OptionsComponentExtensions,
@@ -10,8 +8,6 @@ import type {
 	SpellCheckConfig,
 	TypedFlatConfigItem,
 } from "../types.ts";
-
-const require = createRequire(import.meta.url);
 
 export async function spelling(
 	options: OptionsComponentExtensions & OptionsFiles & OptionsIsInEditor & SpellCheckConfig = {},
@@ -28,17 +24,7 @@ export async function spelling(
 		...componentExtensions.map((extension) => `**/*.${extension}`),
 	];
 
-	const robloxDictionary = require.resolve("@isentinel/dict-roblox");
-	const urlRobloxDictionary = pathToFileURL(robloxDictionary);
-	const urlRoblox = new URL("dict/roblox.txt", urlRobloxDictionary);
-
-	const rbxtsDictionary = require.resolve("@isentinel/dict-rbxts");
-	const urlRbxtsDictionary = pathToFileURL(rbxtsDictionary);
-	const urlRbxts = new URL("dict/rbxts.txt", urlRbxtsDictionary);
-
 	const pluginCspell = await interopDefault(import("@cspell/eslint-plugin"));
-
-	const enabled = inEditor === false ? isInEditor : true;
 
 	return [
 		{
@@ -50,32 +36,7 @@ export async function spelling(
 		{
 			name: "isentinel/spelling",
 			files,
-			rules: {
-				"@cspell/spellchecker": [
-					enabled ? "warn" : "off",
-					{
-						autoFix: false,
-						checkComments: true,
-						cspell: {
-							dictionaries: ["roblox", "rbxts"],
-							dictionaryDefinitions: [
-								{
-									name: "roblox",
-									path: urlRoblox.href,
-								},
-								{
-									name: "rbxts",
-									path: urlRbxts.href,
-								},
-							],
-							language,
-							words: ["isentinel"],
-						},
-						generateSuggestions: isInEditor,
-						numSuggestions: isInEditor ? 8 : 0,
-					},
-				],
-			},
+			rules: spellingRules({ inEditor, isInEditor, language }),
 		},
 	];
 }
