+++ src/rules/jsdoc.ts
@@ -1,0 +1,100 @@
+import type {
+	JsdocOptions,
+	OptionsProjectType,
+	OptionsStylistic,
+	TypedFlatConfigItem,
+} from "../types.ts";
+
+/**
+ * JSDoc rules shared between the ESLint and oxlint factories.
+ *
+ * @param options - Shared rule options.
+ * @returns The rule map.
+ */
+export function jsdocRules({
+	full = false,
+	stylistic = true,
+	type = "game",
+}: JsdocOptions & OptionsProjectType & OptionsStylistic = {}): TypedFlatConfigItem["rules"] {
+	const isPackage = type === "package" || full;
+
+	return {
+		"jsdoc/check-access": "warn",
+		"jsdoc/check-param-names": ["warn", { checkDestructured: false }],
+		"jsdoc/check-property-names": "warn",
+		"jsdoc/check-types": "warn",
+		"jsdoc/empty-tags": "warn",
+		"jsdoc/implements-on-classes": "warn",
+		"jsdoc/informative-docs": "warn",
+		"jsdoc/no-defaults": "warn",
+		"jsdoc/no-types": "warn",
+		"jsdoc/no-undefined-types": "error",
+		"jsdoc/require-description": [
+			"warn",
+			{
+				exemptedBy: [
+					"hidden",
+					"ignore",
+					"inheritdoc",
+					"client",
+					"server",
+					"see",
+					"metadata",
+				],
+			},
+		],
+		"jsdoc/require-description-complete-sentence": "warn",
+		"jsdoc/require-param-description": "warn",
+		"jsdoc/require-param-name": "warn",
+		"jsdoc/require-property": "warn",
+		"jsdoc/require-property-description": "warn",
+		"jsdoc/require-property-name": "warn",
+		"jsdoc/require-rejects": "error",
+		"jsdoc/require-returns-check": "warn",
+		"jsdoc/require-returns-description": "warn",
+		"jsdoc/require-yields-check": "warn",
+		"jsdoc/sort-tags": "off",
+
+		...(isPackage
+			? {
+					"jsdoc/require-param": [
+						"warn",
+						{ checkDestructured: false, exemptedBy: ["ignore"] },
+					],
+					"jsdoc/require-returns": ["warn", { exemptedBy: ["hidden"] }],
+					"jsdoc/require-template": "warn",
+				}
+			: {}),
+
+		...(stylistic !== false
+			? {
+					"jsdoc/check-alignment": "warn",
+					"jsdoc/convert-to-jsdoc-comments": [
+						"warn",
+						{
+							allowedPrefixes: [
+								"@ts-",
+								"istanbul ",
+								"c8 ",
+								"v8 ",
+								"eslint",
+								"jshint",
+								"jslint",
+								"globals",
+								"exported",
+								"jscs",
+								"oxlint-",
+								"prettier-",
+							],
+						},
+					],
+					"jsdoc/multiline-blocks": "warn",
+					"jsdoc/no-blank-block-descriptions": "warn",
+					"jsdoc/no-blank-blocks": "warn",
+					"jsdoc/no-multi-asterisks": "warn",
+					"jsdoc/require-asterisk-prefix": "warn",
+					"jsdoc/require-hyphen-before-param-description": "warn",
+				}
+			: {}),
+	};
+}
