+++ src/rules/roblox.ts
@@ -1,0 +1,42 @@
+import type { OptionsStylistic, TypedFlatConfigItem } from "../types.ts";
+
+/**
+ * Roblox rules shared between the ESLint and oxlint factories.
+ *
+ * These do not require type information; the type-aware roblox rules stay in
+ * the ESLint config (oxlint jsPlugins have no type information).
+ *
+ * @param options - Shared rule options.
+ * @returns The rule map.
+ */
+export function robloxRules({
+	stylistic = true,
+}: OptionsStylistic = {}): TypedFlatConfigItem["rules"] {
+	return {
+		"cease-nonsense/no-array-size-assignment": "error",
+
+		"roblox/no-any": "error",
+		"roblox/no-enum-merging": "error",
+		"roblox/no-export-assignment-let": "error",
+		"roblox/no-for-in": "error",
+		"roblox/no-function-expression-name": "error",
+		"roblox/no-get-set": "error",
+		"roblox/no-implicit-self": "error",
+		"roblox/no-invalid-identifier": "error",
+		"roblox/no-namespace-merging": "error",
+		"roblox/no-null": "error",
+		"roblox/no-private-identifier": "error",
+		"roblox/no-unsupported-syntax": "error",
+		"roblox/no-user-defined-lua-tuple": "error",
+		"roblox/no-value-typeof": "error",
+		"roblox/prefer-get-players": "error",
+		"roblox/prefer-task-library": "error",
+
+		...(stylistic !== false
+			? {
+					"cease-nonsense/no-array-constructor-elements": "error",
+					"sentinel/prefer-math-min-max": "error",
+				}
+			: {}),
+	};
+}
